create password utils.py
